[PERF] planning_timesheet_forecast: speed up unlink of `planning.slot`
Description:
Deleting planning slots can be slow on some databases that make heavy
usage of the Timesheet/WorkEntries from shifts feature and/or they have
large tables for their `account.analytic.line` and/or `hr.work.
entry`. It can be especially felt when deleting all recurrences of a
shifts.
This is due to the missing indexes on the FKey to the
`planning.slot` model on large models. Without an index on the FKey,
deleting a `planning.slot` will trigger Seq.Scans on the tables
that references the `planning.slot` model to set the field to NULL
or trigger the SQL `ON DELETE` constrains.

Fix:
Add the missing indexes. Partial on nulls because it's a sparse
relationship.

Reference:
task-3747131

closes odoo/enterprise#57942

X-original-commit: 5c40cbf1f0035088ac87a2683d61e17bf4a2a218
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>