[FIX] account_intrastat: Fix warning traceback
During a forward port (https://github.com/odoo/enterprise/commit/bb67df2f37a65acd9f62a28d8a52fbd335ef903e),
we introduced a bug because the old warning system
was in conflict with the new one.
The aim of this commit is removing the old one (that
are not working) for the new one. In the same time,
we make sure that we use the right key when getting ids
that have an issue.

opw-3741287
opw-3741721
opw-3741658

closes odoo/enterprise#56738

X-original-commit: 26fc172004d880f5e941c78cbdc18a29d2db5134
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>