[FIX] website_helpdesk_livechat: chatbot steps validation with create ticket
Before this commit:

When you have a `Create ticket` step in your ChatBot script steps,
but you don't have an `Email` step prior to that.
Then it will break the flow of ChatBot, as an email is required to
create a helpdesk ticket.

After this commit:

It will raise a `ValidationError` if it fails to validate the steps
while configuring the chatbot script steps.

task-3547675

closes odoo/enterprise#64430

X-original-commit: 0318366749f88c8effa2bbad15c07d160b41b71f
Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>
Signed-off-by: Rahul Prajapati (rapr) <rapr@odoo.com>