[FIX] account_accountant: remove useless test and tour
The tour is pretty useless and has been improved in 17.0.
We will keep it in that version, to be sure that we don't break
the equivalent tour, but we can remove in these versions (it just creates
a sample bill and post it, currently).

The tour `account_accountant_tour_upload_ocr_step` has been deemed
useless. It pops a bubble on a button where it makes no sense, with
a self-explanatory message already.

closes odoo/enterprise#57225

X-original-commit: efe4f4517ae9f2fbd69e09c4ef2c04e502f5b9f5
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>