[FIX] account_accountant: remove useless test and tour
The tour is pretty useless and has been improved in 17.0.
We will keep it in that version, to be sure that we don't break
the equivalent tour, but we can remove in these versions (it just creates
a sample bill and post it, currently).

The tour `account_accountant_tour_upload_ocr_step` has been deemed
useless. It pops a bubble on a button where it makes no sense, with
a self-explanatory message already.

closes odoo/enterprise#57414

X-original-commit: 1b6811fbcd8bbcb0f33b82fcb076215b378684cf
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>