[IMP] knowledge: allow item creation from opened embed views
Purpose:
--------
Currently, users cannot create items from opened embed views (except for
the kanban view from which it is possible to use the quickcreate feature
regardless of the user's right on the parent article, which may lead to
access errors).

The limitation is due to the usage of the environment to store the user's
write permission on the parent article, which is lost when opening the view.

Instead of using the environment, this piece of data is now stored in the
context, and is therefore set when the view is embedded and when it is
opened in full screen.
When the opened view is reloaded, the context is lost and the actions are
enabled by default, except for
- the quick creation in the kanban views as other keys needed for it to
work as expected are missing
- the calendar view, as it is in "missing configuration" mode in that case

Task-3607081

Part-of: odoo/enterprise#56706