[FIX] appointment: fix unpublished image access
Purpose
=======
Fix the unpublished appointment types images which weren't showing for
public users when accessing the invitation link. Instead of the
appointment type image, the default web placeholder was displayed.

Specifications
==============
As the appointment types are unpublished and the user is public,
when trying to retrieve the image by using `ir.binary._find_record`
it was checking the access rights/rules on 'read' and launching an AccessError.
As the record wasn't found, a fallback was made on the default web placeholder image.

Fixing that by always granting the access to the appointment type image.

Task-3725132

closes odoo/enterprise#58031

X-original-commit: 4e2a21b25ed5d0e1f8c28d39f2d65b85f8433ec4
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>