[IMP] l10n_lu_hr_payroll: Move _get_lux_tax method to rule data
Purpose
=======

Even if the end_user is a tech, that he knows the code, as long as
it's on the saas, he just can't do anything.

Calling method in the rules is counter intuitive, you think that you
have the hand, can check a parameter or not, but in fact you can't do
anything.

closes odoo/enterprise#56710

Taskid: 3748649
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>