[FIX] hr_contract_salary: disable dependent benefits
When we unselect the mandatory benefit, the dependent
benefit should be unset and disabled.
But before this commit this is not a case.

The explanation of an issue:
The issue only happens if dependent benefit has
value different from 0, so the value should be unset first
before disabling the dependent benefit.
When the dependent benefit should be unset, it
calls the asynchronous function, which prevents
the lines that disable the dependent benefit to be reached.

task-3749171

closes odoo/enterprise#56824

X-original-commit: 4525a8c764920512b468988de985f10522f2158a
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>