[FIX] hr_contract_salary: disable dependent benefits
When we unselect the mandatory benefit, the dependent
benefit should be unset and disabled.
But before this commit this is not a case.

The explanation of an issue:
The issue only happens if dependent benefit has
value different from 0, so the value should be unset first
before disabling the dependent benefit.
When the dependent benefit should be unset, it
calls the asynchronous function, which prevents
the lines that disable the dependent benefit to be reached.

task-3749171

closes odoo/enterprise#57088

X-original-commit: 70e60e261effa1d9f1f31d9b03c9e9d9241578e5
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>