[FIX] helpedesk : search by assignees with portal user
Steps to reproduce:
	1- Install Helpdesk, Website module
	2- Login using portal user
	3- Navigate to tickest and search by assignees

Current behavior before PR:
When searching by assignees using a portal user you will not get any results. This is happening because the portal user seems to not have the access to search in 'res.users' so when the orm parse the user_id assigned to the ticket to a name it will not be able to parse it for the portal user.

Desired behavior after PR is merged:
We search first for the users that are matching the search key in 'res.users' using 'sudo()' then get their ids and search in 'helpdesk.ticket' using those ids

opw-3719387

closes odoo/enterprise#57757

X-original-commit: e2b7bd095f2e4f5971f22c323ad83f0cb2a3e484
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Youssef Bashandy (yoba) <yoba@odoo.com>