[I18N] l10n_fr_reports: restructure po/pot files
Add the missing terms to the pot file.
When the terms are missing from the pot file, but present in the po
file, they get marked as "obsolete".
This happens during the merge function that is called within
tools/translate.py.

In order to avoid this, the pot and po files have been re-generated and
restructured in order to make future additions easier to merge.

closes odoo/enterprise#57905

X-original-commit: 13f2965e548baa9fd67ac5cf9e38de657954401c
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Daniel Kosky (dako) <dako@odoo.com>