[FIX] project_enterprise: remove unnecessary step from tour
Versions:
---------
- saas-17.1

Issue:
------
The test tour is failing.

Cause:
------
The 'allocated_hours' step was added in the test tour, even though the 'View Task'
form view does not have this field until timesheet is enabled.

Solution:
---------
Removing this step as the 'allocated_hours' field is not present in the view.

task-3750478

closes odoo/enterprise#57169

X-original-commit: d0c2a49b6d75d4472509105b9abb253daa58113e
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>