[FIX] web_studio: report: copy all translations from all inheritance
Have some translations set on a view used by a report.
Edit that report with studio.

Print the report with a specific language.

Before this commit, the terms that were translated before the edition, were not translated
after edition.

Since commit[1], the whole combined arch of a view is used to make a final inheriting view
containing the changes. Unfortunately, translations were in the blind spot of that commit, so that
there weren't lost, but simply not copied on the studio view.

After this commit, this works as expected as we copy all terms for all languages for all views
in the inheritance chain onto the studio view.

opw-3746922
opw-3724701
opw-3719157

[1]: b7d0d3034d53cdab831f42f8b37a14cf50f3720e

closes odoo/enterprise#57490

X-original-commit: 7334697dcb869b61594ebc24266a415c0e1b212d
Related: odoo/odoo#155346
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>