[FIX] account_reports: Wrong auditing on aged reports
This commit modifies the behavior of the auditing of the aged reports.

When auditing the aged receivable for example, we would only get the amls from
purchase journals. We also want the misc and bank for example so while we
previously would check that journal is equal to the the journal_type provided,
we now exclude the opposite journal_type

Also the date_scope of the report was badly passed into the domain. This could
lead to elements not being shown

task-3640027

Part-of: odoo/enterprise#54137