[FIX] l10n_ar_edi: Fix memory issue with computation of field
Issue -->

For a database with a large number of `account.move` records, the computation of the field
`l10n_ar_fce_transmission_type` leads to a MemoryError when installing the `l10n_ar_edi` module.

Solution -->

Use the `_auto_init` method to add the column to the `account_move` table to skip computation during
install.
Set the value to NULL on init because the field does not have a default value.

opw-3694393

closes odoo/enterprise#57398

X-original-commit: d66709bbee6e78509e93ef9e7f38427c29d89fd9
Signed-off-by: Cedric Snauwaert <csn@odoo.com>