[FIX] sale_amazon: set stock move as done in FBA
Currently, synchronizing done order in FBA creates some stock move
that are not in the state done anymore.

This is due to some changes in the management of the quantity done in
stock, that needs to see if the quantity was picked or not. As, in
this case, we are confirming a move for products that were stocked in
an Amazon Warehouse and then delivered by Amazon, we can manually set
the move as being picked.

opw-3594556
opw-3662990

closes odoo/enterprise#56943

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>