[FIX] helpdesk_timesheet: ensure test use data created in test case
Before this commit, the ticket is set to a random helpdesk team with the
id equals to 2 which is surely the id of a demo data (VIP Support).

This commit makes sure the ticket created inside the test will be linked
to a helpdesk team created in the test and not a helpdesk team linked
to the demo data.

runbot-57412

closes odoo/enterprise#56990

X-original-commit: fdbfd4b85c56239f1594fcd3c799b6b50df9e8fb
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>