[FIX] web_studio: approvals not colliding when multiple model have same method
Have two models that have a method with the same name.

Create multiple rules for model 1 and 1 one rule for model 2

Before this commit, the rule on model 2 will interfere with model 1.
This was because some domains missed to filter on the model of the rule

After this commit, rules are not colliding between models.

opw-3734028

closes odoo/enterprise#57082

X-original-commit: 7c8c61f63af701a7f387362438cdc13aba674cd7
Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>