[ADD] documents_account_peppol: receive peppol documents in documents
Currently we are able to receive Peppol documents in a specific journal.

When working with multiple purchase journals, or having specific flows
in the Documents app to process incoming vendor bills, it makes sense to
receive your Peppol documents in the Documents app to process them.

This change does just that by giving the user the option to receive
incoming Peppol documents either in a journal, the Documents app, or
both. It also allows to assign specific tags to the document.

task-3519604

closes odoo/enterprise#61606

X-original-commit: ef3f371c6eec3fa2d58e42b8619ee1984d175085
Related: odoo/odoo#163632
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>