[FIX] l10n_mx_edi: set correct payment way
Summary
-------
Currently, when creating a payment, the "payment way" defined on the
customer takes precedence over the payments specified on the payment
wizard.

Steps to reproduce
------------------
* install  `l10n_mx_edi`
* switch to a Mexican company
* create a contact C
* on the contact form, in the "Sales & Purchase" tab, set the "Usage"
  and "Payment Way" to "General expenses" and "Tarjeta de Crédito"
  respectively.
* create an invoice for contact C, with a due date for next month
* register a payment, with "Payment way" set to "Efectivo"

The payment that is created will have it's  "Payment way" set to
"Tarjeta de Crédito" (the one defined on the customer), even though you
specifically selected "Efectivo" in the payment wizard.

opw-3717693

closes odoo/enterprise#55850

Signed-off-by: Laurent Smet (las) <las@odoo.com>