[FIX] appointment: ignore calendar for appointment resource leave
Make the resources field required in the view. We also hide
the calendar field when creating a leave for an appointment
resource. Keeping this field and the resource empty was leading
to the possible creation of general public leaves for every type
of resources (even users). It is now hidden as it does not make sense
to keep it displayed when it will be ignored because of the required
resources.

The goal is to remove and clean complety the resource calendar field
from this view/model later.

X-original-commit: d37a3bf03891e9fcc0016add132f2971b33213c8

task-3562191

closes odoo/enterprise#57438

X-original-commit: 9cbbf8b4e6e0d5e3e6599234971d4649d98a52d1
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>