[FIX] account_intrastat: Make sure we have unique generic id
The aim of this commit is making sure that we don't have
inconstancies between the GROUP BY in the _build_query_group
and the intrastat report line generic id.
Indeed, before this commit, the incoterm code, the transport code
and the invoice_currency_id were not included in the generic id.
It means that we can have several lines with the same generic id
but with different values (for example incoterm code).
This commit adds a test that check that all elements in the GROUP BY
(except the system, type and region code) are used in the generation
of the generic id. For the excepted keys, system and type are already
separated. Concerning the region_code, its value related to the company,
it means that this value is not discriminant.

opw-3741658

closes odoo/enterprise#57272

X-original-commit: 41ec084783355910972da57d63ebef7087c06ad3
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>