[FIX] account_followup: unblock the search on 'followup_status'
Problem: When a user sets a custom filter on field 'followup_status'
to "is set/is not set", a traceback error occurs with the message:
"TypeError: 'bool' object is not iterable".

Solution: To prevent the traceback error, the check for value should be
done first.

Steps to reproduce on Runbot:
1. Install Contacts and Accounting
2. Navigate to Contacts and add a custom filter for
"Follow-up Status is set/not set" and traceback error occurs

opw-3736643

closes odoo/enterprise#57209

X-original-commit: 2afcea7c129148e9be158adb7728888525547cd4
Signed-off-by: Mylyna Hy (myhy) <myhy@odoo.com>
Signed-off-by: Cedric Snauwaert <csn@odoo.com>