[IMP] l10n_au_hr_payroll: sequence improvement
With termination payments, we will need to add
additional rules on top of the rules in the
various schedules so that we only end up with
a single payslip.

To that end, we will rework a bit the sequences
in the other schedules in order to avoid having
all numbers following each others and to leave gaps
that makes it easier to insert rules at different
stages when needed.
This also aims to make the same rules have the
same sequence in the different structure, as well
as avoid having two rules with the same sequence.

We will also rearrange the file so that rules
appears in the file following their sequences,
to make it easier to process how the lines
will look.

Finally, we add xml comments where the standard
rules will appear to make it extra clear.

Task id # 3639270

Part-of: odoo/enterprise#57884