[FIX] l10n_es_reports: some modelo 347 fixes
Purpose:
Fixes applied to modelo 347 based on feedback (Task: 3757213).
Addressing issues with VAT ID being included incorrectly and default
invoice type for foreign partners as specified in the AEAT guidelines.

Specs:
1. Corrected the default computation of the invoice field in the AEAT
tab for non-EU partners, ensuring it remains empty as per specifications
2. Modified the BOE export logic to prevent duplicating the VAT number.
Ensured that for Spanish partners, the VAT ID is placed in
'NIF Declarado' and for EU (non-Spanish) partners,
in 'NIF OPERADOR COMUNITARIO', aligning with the guidelines.

task-3758760

closes odoo/enterprise#57470

X-original-commit: bf7bff5f857bc27f7322692ef48ab947a1f91a28
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Mohamed Erradi (moer) <moer@odoo.com>