[IMP] account_reports: reconciliation reports
In this https://github.com/odoo/enterprise/commit/76caf5f5675f97128b88fbcffedaf1e85da516d5,
it was decided to remove the report. But we received a lot of feedbacks about
users missing the report, especially from anglo-saxon countries, where this
report is often used in Audit processes to report on the differences between the
Accounting and Bank reported amounts on bank accounts.

In this pr we decided to put it back and improve it. What has been improved:
- The display of the report is now consistent with the different reports
- Use of custom engine
- New sections added
- Rework of the warnings

closes odoo/enterprise#58640

Task: 3756179
Related: odoo/odoo#157648
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>