[FIX] account_reports: handle duplicate aml id in partner ledger
The owl key used was the id of the journal item, it is not unique in some
cases of indirect linked aml (amls which don't have the partner of the report
section but are reconciled with an aml which have this partner).

An indirect aml can have multiple partial reconciliations with different amls
from the same partner. Also the way odoo shows indirect aml in "Unknown partner"
section is to duplicate all info shown except debit/credit/balance from origin
aml (so this include aml.id).

To support those cases, we now use a markup with the id of partial reconcile to
produce unique keys. When the aml is not an indirect aml, this markup is set to
0.

Steps to reproduce:
1/ Create an invoice with partner
2/ Create a MISC entry without partner for the same amount
3/ Reconcile both
4/ Go to partner ledger for the right period + unfold all => Traceback

opw-3700478

closes odoo/enterprise#58212

X-original-commit: 9e3a22a77247ff2281f77a3f7e4a0f7c6005e7d9
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Thomas Beckers (tbs) <tbs@odoo.com>