[FIX] web_studio,website_studio: fix adding a website form
In the website tabl, try to add a website form
Before this commit, there was a crash because of some callback that was not bound.
This was caused by commit e3c8f97d0b8553abc1e77bd3699fcb880e48ab0a

After this commit, the flow works as expected.

opw-3735624

closes odoo/enterprise#57155

X-original-commit: f8b96474f7b461ce3913183cae5219a75b8ba9a1
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>