[FIX] account_intrastat: 2024 add/remove codes
The annual changes to the intrastat code lists have occured. In order to
synchronise with these new standards, we have to add the new codes, and
remove the old ones.

This commit adds the new codes in into the relevant csv. We also add
expiry dates for the removed codes, and start dates for the new codes.
In forward ports, these should be added directly on the
account.intrastat.code.csv (since the modules account_intrastat and
account_intrastat_expiry were merged after v15.0)

One of the codes (50072060) has a supplementary unit, so we add that to
the relevant dictionary, though in the forward ports this will be added
directly in the data.

closes odoo/enterprise#57890

Task-id: 3747808
X-original-commit: 66503d0a48feb552dc726600c7c12523954d7c88
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Daniel Kosky (dako) <dako@odoo.com>