[IMP] l10n_ke_hr_payroll: Added Master Report
This commit adds the master report to the payroll
module. The master report is a report that shows
relevant information about the payslips of a given
period. The report is accessible from the payroll
report menu.

This commit also adds some utils for all the
reports with the same kind of structures.

task-3744470

closes odoo/enterprise#57133

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>