[FIX] account_accountant: match exact amount when text/partner check fails
Create a reconciliation model as follows:
- Type: Rule to match invoices/bills
- Match Invoice/bill with Label: False
- Match Invoice/bill with Note: False
- Match Invoice/bill with Reference: False
- Partner is set: False
Create an invoice for $900 and confirm it.
Create a bank statement line for $900, with no partner set.
Go to reconcile the bank statement line.

Issue: The invoice is not automatically suggested

This occurs because there is no check for the exact amount as in
previous version

opw-3648260

closes odoo/enterprise#57406

X-original-commit: ce47035c3b7dc1a4bf4d798cef93ad4cfbf0b747
Signed-off-by: Laurent Smet (las) <las@odoo.com>