[FIX] sale_subscription: report subscriptions with billing period of 0
Steps to reproduce:

- Subscriptions > Configuration > Reccurence periods.
- Pick Monthly and set duration to zero.
- Subscriptions > Reporting > subscriptions

 **Traceback Error**

Cause of the issue:

The SQL queries:
https://github.com/odoo/enterprise/blob/f673997a361e97ff2a22d92f9bf34e7277992039/sale_subscription/report/sale_subscription_report.py#L50-L72
divide by zero when the `billing_period_value` field of the
`sale.subscription.plan` model is set to 0.

Fix:

Since the `billing_period_value` field of the `sale.subscription.plan`
model should not be allowed to be set at 0 see: commit 194ab0f
For our fix to be stable, we propose here to implement an ORM constraint
that will somewhat achieve the same duty as the sql constraint
introduced in 17.1: commit 48aef7d but without breaking existing DB's
that already uses incorrectly the `billing_period_value` field.

Note:

This fix ensures that the issue will not be reproducible anymore but
does not solve the issue for DB's that already uses incorrectly the
`billing_period_value` field. For the subscriptions reports to work
correctly on these DB's, the `billing_period_value` will still need to
be set manually to acceptable values on each `sale.subscription.plan`.

opw-3733854

closes odoo/enterprise#57165

Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>