[REF] web: remove include archived toggle from documents model
Since we can no longer archive documents, the 'Include Archived' toggle checkbox
of the domain selector is now irrelevant and hence this commit removes it.

Technical:
The archive feature for documents has been removed with #37389.
However, the active field is still being utilized for the 'trash' feature
and hence cannot be removed.

Whether the `Include Archived` toggle should be displayed or not,
is determined by checking if a particular model has 'active' in its fieldDefs
(i.e. if the model has archive/unarchive feature).
As a result, the condition becomes true in case of 'documents' model
despite archive records feature not being available anymore.

Thus, this commit adds a condition to check resModel, and hides the toggle
checkbox from the domain selector in case of 'documents' model, by deleting
'active' from fieldDefs.

Task: 3695462
Part-of: odoo/enterprise#57208