[FIX] web_studio: include hoisted fields in kanban
This commit fixes the way we filter fields to show in the kanban
editor sidebar, under the 'Add' tab. Since this view don't render
what's included before the <templates> node from the xmlDoc.

In other types of view, we can easily toggle the 'Show invisible'
elements to display items that are considered being present in the
view, but not in Kanban when they add items before the kanban card
template.

A test has been added too.

task-3647045

closes odoo/enterprise#58911

X-original-commit: c35edf32565b4a18fcca07ee5c425df89902ceb6
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>