[IMP] web_grid: adapt cell height and highlight
We were not seeing enough data on a low-res screen and the column
highlight was too distracting (fp request).

This commit reduces the cell height but keeps it big on mobile to ease
finger selection. The column highlight is removed.

Since we do not highlight the column anymore and the cell height is
smaller. It also removes the cell highlight on add a line and the dates,
thus we have to adapt the tests accordingly.

The cell highlight is also slightly less gray.

task-3751591

closes odoo/enterprise#57257

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>