[FIX] voip: no crash on transfer to another device
The forward-port of #54535 neglected to adapt the code it introduced to
the move of settings to its own model. Because of this oversight,
this.settings is undefined, and this.settings.external_device_number
would crash.

This commit adapts the code to properly retrieve the config, and adds a
test to make sure this never happens again.

opw-3743892

closes odoo/enterprise#57294

X-original-commit: 984c2d964594c36dd8d9ad909cadbacd5ba1e89f
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>