[FIX] stock_barcode_picking_batch: hide confirm button in barcode view
To reproduce issue:
.go to barcode app
.click on batch tr
.click on new
.click on engine icon

Expected result:

confirm button hide

Actual result:
confirm button is not hide

When a user navigate from batch transfer and click on a engine icon
the confirm button still showing, it's no makes sense for the current
view. Now, when the user click on it, the confirm button will be hide
and apear when close that menu.

task-3525776

closes odoo/enterprise#57163

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>