[IMP] planning(_contract): take unavailabilities into account in recurrency
This commit introduces unavailabilities when planning recurrent shifts in
the planning app. Unavailabilities include company closing days, public
holidays, employee leaves, shifts in conflicts, fixed contract endings, etc.

Prior to this change, when recurrent shifts were planned and generated, no
unavailability information was taken into account. For example, if we planned
a shift to repeat every month on the 15th, and it landed on a weekend (a
company closing day), it would be normally generated. After this commit, it
would not be, as no resource is supposed to be working at that time.

Also, changed the way we generate the slot when the "x_times" repeat type is chosen.
Now the next_shift method is changed to yield only generate-able shifts plus
it keeps track of repetitions instead of looking at the range limit.

task-2918840

closes odoo/enterprise#57356

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>