[FIX] account_intrastat: Make sure all values are in company currency
Before this commit, values in grouped lines are not in the
company currency but in the foreign currency set on the move.
This is wrong, this commit changes the way we're computing values to
get the value in company currency instead of in foreign currency.
To do that, we change the computation (quantity * price_unit)
for this:

Sign of quantity * Sign of price unit * ABS(balance)
It means that we handle possibility that quantity and price unit could be
negative (for some strange corner cases), and we get the amount
in company currency instead of in the foreign currency.

opw-3757444

closes odoo/enterprise#57465

X-original-commit: 2c35627debee2204a99ab856362c36eb2f136b6c
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>