[FIX] worksheet: fix company error message
This commit's purpose is to improve the error message when a user is
adding new company to a template that had none set before.

Steps to reproduce:

- open fsm app
- open configuration menu, worksheet template
- select the 'Device installation and maintenance' worksheet (it is the
one with tasks linked to it)
- add a new company 'my company (chicago)' and save.

A validation error message pops up. That is fine and the expected
behavior, since the new companies of the template are not consistent
with the task already linked to it.

Source:
the generic message is using the old values of the companies of the
worksheet, but since there were none, no names are displayed and the
error message is confusing.

Solution:
check if the worksheet used to have companies set before the user
changes. If not, display a relevant error message

task-3749225

closes odoo/enterprise#59931

X-original-commit: fce130e62d205a72838bd59457e83c0107852834
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>