[FIX] delivery_ups_rest: include `street2` in address line
Steps to reproduce:
create a shipment to a customer and validate it. `street2` field
is not included in the address line sent to UPS and it does not
appear on the label.

The issue happens because in line
https://github.com/odoo/enterprise/blob/bec5f384325116e21221260559c7141cccf0500a/delivery_ups_rest/models/ups_request.py#L248
if `street` exists, the rest of the line is short-circuited, and
so we will only have that line.

This commit changes the format to one similar to legacy UPS module.

opw-3745075

closes odoo/enterprise#57469

X-original-commit: e7f36b4e996c047e5841367c13df8f24db5c419f
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Ahmad Khanalizadeh (khah) <khah@odoo.com>