[FIX] documents: fix search pannel padding issue
Technical
====================
Due to a recent change here https://github.com/odoo/odoo/pull/149246 there is
padding issue because in documents we are overriding the carets and trash
icons and width of the "o_toggle_fold" is fixed so that's why it is causing the
issue.

After this commit
==================
Removed the unnecessary padding.

Task-3761896

closes odoo/enterprise#57489

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>