[FIX] knowledge: add missing data hotkey for wysiwyg helper
When the user presses the `Alt` key in the Knowledge editor, there is no
hotkey assigned to the button "Generate an Article with AI" located in the
wysiwyg helper. This commit addresses the issue by defining a hotkey for
that button. For that button, we will use the "G" key to refer to "Generate".

task-3777661

closes odoo/enterprise#58174

X-original-commit: 035113b4b542240488767d750f0c4abf50bef796
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>