[FIX] appointment: fix appointment types fetching
Fix the fetching of the appointment types which was triggering
an Access Error for public and portal users as they didn't have
the read access on the appointment.type model.

Instead of calling the search method directly without sudo to apply
the user access rules, rolling back to the 17.0 version where the search was
performed using a sudo and the resulting recordset was then filtered
based on the user access rules. This guarantees an access to the
appointment.type model and always returns the appropriate recordset.

related commit: 05a1080a66cefc1b0724e23c72dd2bf06343e4e3

Task-3765016

closes odoo/enterprise#57767

X-original-commit: 9d8bc3830142a564909a5680a3680649bfcd5012
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>