[FIX] mrp_workorder: fix unscoped change on the modals
This commit fixes an issue about an overflow property being applied to
all the modal across Odoo, while it should only affect a single module.

During the design and UX improvements of `mrp_workorder`, Commit[1]
introduced an overflow property on the `modal-body` in order to fix a
layout on modal that have hidden content that is meant to be displayed
when clicking on a button.

While this line provided a fix for the issue, it was not scoped to the
module and would then affects the whole back-end.

To fix this issue, we add the property as a utility class on the specific
modal.

Commit[1]: 6f9b089

ticket-3671725

closes odoo/enterprise#57663

X-original-commit: ce0c43ff80b4197e0a1806e8847080648a61848c
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Chrysanthe Gomrée (chgo) <chgo@odoo.com>