[FIX] planning: ensure split tool is correctly placed
Before this commit, when the gantt view displays the current week,
let's say from `2024-18-02` to `2024-24-02` and the planned dates of one
pill is:
- `start_datetime`: `2024-17-02 16:00:00`
- `end_datetime`: `2024-18-02 00:00:00`
(the end datetime of the pill should be the first column displayed in
the gantt view), a traceback is occurred because the pill will not be
displayed in the gantt view and so the grid column for that pill will be
0 and since the split tool uses the grid column for each pill to place
the different scissor to enable the split tool for the pills displayed
on more than one column in the gantt view.

This commit makes sure the split tool will skip the pill with column
less than 1 to be sure `pill.grid.column[0] - 1` uses to create a array
will not be less than 0.

opw-3705995

closes #55807

closes odoo/enterprise#57509

X-original-commit: d8c76685d09e724ca2060b7dcbcbc062e1fa7444
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>