[FIX] industry_fsm_sale: compute catalogue from tasks instead of main SO
Steps to reproduce:

- Create an SO with two lines containing the product “Field Service”
- Confirm the SO > 2 fsm tasks are created
- Open the first task > access catalog via the `Products` smartbutton
- Add any product to the SO.
- Open the second task > access catalog via the `Products` smartbutton
- Add the same product to the SO.
- Open the first task > access catalog via the `Products` smartbutton
- Try to add more of that product

> IMPOSSIBLE : “You can’t edit this product in the catalog”

Cause of the issue:

The way the catalogue is computed was refactored between 16.4 and 17.0
see odoo commit bc01c7bcec974ee095f5855f225667645ef40213
The above situtation is not yet handled by this refactoring.
In 16.4, since the catalogue was acceced from the task linked to the
Field service, the quantities it displayed were computed with respect to
that task. However, in 17.0 the quantities appearing in the catalogue
are computed from the main SO containing both `Field services` SOL
instead of being computed contextually from the task they are linked to.

To be more precise:

When clicking on the smart button, the quantities appearing in the
calatalogue are computed by the `_get_product_catalog_order_line_info`.
This method starts by grouping the SOL referencing each product of the
catalogue using the `_get_product_catalog_record_lines`:
https://github.com/odoo/odoo/blob/c7f982774654459a2db6decd6ae1948e05e7d622/addons/product/models/product_catalog_mixin.py#L91-L92
https://github.com/odoo/odoo/blob/c7f982774654459a2db6decd6ae1948e05e7d622/addons/sale/models/sale_order.py#L1823-L1829
But here is the problem, since both SOL were able to reference the
same product of the catalogue the product of the catalog will be set to
'readOnly' here (because `self` will contain both SOL):
https://github.com/odoo/odoo/blob/c91ffa3bdbff845088111bc3036500c2c03fc357/addons/sale/models/sale_order_line.py#L1222-L1227
Hence, we will not be able to edit its value from the catalogue:
https://github.com/odoo/odoo/blob/c91ffa3bdbff845088111bc3036500c2c03fc357/addons/product/static/src/product_catalog/order_line/order_line.xml#L18-L20

Fix:

As discussed with the PO of the Field Service module (see conversation
of the ticket), the catalogue should be computed contextually using the
task from which we access it and not from all tasks referring to that
product in the main SO. Doing so will solve the problem since only one
SOL will be associated to that product of the catalogue and the product
will be set to `'readOnly' : False`:
https://github.com/odoo/odoo/blob/c91ffa3bdbff845088111bc3036500c2c03fc357/addons/sale/models/sale_order_line.py#L1212-L1216

opw-3748833

closes odoo/enterprise#58385

X-original-commit: 8142017b2c0388be03242431bb46b57ecbec4caa
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>