[FIX] helpdesk_sale_timesheet: put team id in the context
Steps to reproduce:
-------------------
- go to Sales;
- click on a sale order with tickets assigned to it;
- click on the "Tickets" smart button;
- group by stage.

Issue:
------
A traceback appears.

Cause:
------
The query formed is incorrect.
We are using a Query object in the domain instead of the team id.
Which has been placed in context beforehand.

Solution:
---------
Place the team id in the context.

Note:
-----
This will automatically be converted to a list if necessary
when generating the query with a domain using the `IN` operator
(relation).

opw-3759882

closes odoo/enterprise#57502

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>