[FIX] account_invoice_extract: incorrect automatic sending from alias
Since commit d79ac56, invoices coming from the email alias were
automatically sent to the OCR even when the automatic sending mode was
disabled.

The logic that decides whether or not the document should be
automatically sent to the OCR was split across multiple places, making
this kind of error easy to make.
This commit addresses the issue by regrouping all of this logic inside
the `_needs_auto_extract` function.

Ticket #3708420

closes odoo/enterprise#57553

X-original-commit: 519ce5f92c2889413b6169b1b7abce62032cdced
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>