[IMP] hr_referral: UX improvements
Makes the referral cards take the full width of the screen,
improve the looks of the button by the use of logos and
update the twitter logo with x
Note: The twitter logo is not in the fa library, that we use
so I had to import it.

This commit remove completly the use of the referral
link to share wizard for this and embed the link directly in
the button (that was opening the referal link to share)
to copy it directly to the clipboard.

This commit adds the ability to send a referral
link via SMS.
This commit also put the new relation inside the
send mail wizard.

task-3418434

closes odoo/enterprise#53037

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Co-authored-by: mepe-odoo mepe@odoo.com